requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
PyYAML>=6.0
Pillow>=8.0.0 